refactor: move user actions to a separate file
